feat: use object oriented design
The algorithm for generating a secret number, reading the user input,
checking whether the input matches the secret, and saying good bye, was
split into several small classes, so that this program can be considered
object oriented.
Secret, Guess, Diss, VerboseDiff, Attempts and Farewell, are grouped
together to form the main program.